Fix return size bug in lossless compressors & add Zstd lossless compressor

+++ include/mgard-x/Utilities/ErrorCalculator.h
@@ -15,7 +15,7 @@ namespace mgard_x {
 template <typename T>
 T L_inf_norm(size_t n, T * data) {
   T L_inf = 0;
-  for (int i = 0; i < n; ++i) {
+  for (size_t i = 0; i < n; ++i) {
     T temp = fabs(data[i]);
     if (temp > L_inf)
       L_inf = temp;
@@ -26,7 +26,7 @@ T L_inf_norm(size_t n, T * data) {
 template <typename T>
 T L_2_norm(size_t n, T * data) {
   T L_2 = 0;
-  for (int i = 0; i < n; ++i) {
+  for (size_t i = 0; i < n; ++i) {
     T temp = fabs(data[i]);
       L_2 += temp * temp;
   }
@@ -37,7 +37,7 @@ T L_2_norm(size_t n, T * data) {
 template <typename T>
 T L_inf_error(size_t n, T * original_data, T * decompressed_data, enum error_bound_type mode) {
   T error_L_inf_norm = 0;
-  for (int i = 0; i < n; ++i) {
+  for (size_t i = 0; i < n; ++i) {
     T temp = fabs(original_data[i] - decompressed_data[i]);
     if (temp > error_L_inf_norm)
       error_L_inf_norm = temp;
@@ -54,7 +54,7 @@ T L_inf_error(size_t n, T * original_data, T * decompressed_data, enum error_bou
 template <typename T>
 T L_2_error(size_t n, T * original_data, T * decompressed_data, enum error_bound_type mode) {
   T error_L_2_norm = 0;
-  for (int i = 0; i < n; ++i) {
+  for (size_t i = 0; i < n; ++i) {
     T temp = fabs(original_data[i] - decompressed_data[i]);
     error_L_2_norm += temp * temp;
   }
@@ -70,7 +70,7 @@ T L_2_error(size_t n, T * original_data, T * decompressed_data, enum error_bound
 template <typename T>
 T MSE(size_t n, T * original_data, T * decompressed_data) {
   T mse = 0;
-  for (int i = 0; i < n; ++i) {
+  for (size_t i = 0; i < n; ++i) {
     T temp = fabs(original_data[i] - decompressed_data[i]);
     mse += temp * temp;
   }
@@ -81,7 +81,7 @@ template <typename T>
 T PSNR(size_t n, T * original_data, T * decompressed_data) {
   T mse = MSE(n, original_data, decompressed_data);
   T max = 0, min = std::numeric_limits<T>::max();
-  for (int i = 0; i < n; ++i) {
+  for (size_t i = 0; i < n; ++i) {
     if (max < original_data[i]) max = original_data[i];
     if (min > original_data[i]) min = original_data[i];
   }
